+++ planck-cljs/src/planck/repl.cljs
@@ -1026,7 +1026,7 @@
                             (concat (->> requires
                                       (filter #(string/starts-with? % "goog."))
                                       (map (comp goog-dep-source symbol)))
-                              [(first (js/PLANCK_LOAD file))])))
+                              [(first (js/PLANCK_READ_FILE file))])))
                   (deps/js-libs-to-load name))]
     (cb {:lang :js
          :source (string/join "\n" sources)})