+++ planck-cljs/test/deps.cljs
@@ -1,0 +1,1 @@
+{:libs ["planck-cljs/test/libs/mylib.js"]}