+++ planck-cljs/test/general/closure_libs_test.cljs
@@ -1,0 +1,7 @@
+(ns general.closure-libs-test
+  (:require
+   [cljs.test :refer-macros [deftest is]]
+   [my-lib.core]))
+
+(deftest closure-lib-loaded
+  (is (= 5 (my-lib.core/add 2 3))))