+++ planck-cljs/test/libs/mylib.js
@@ -1,0 +1,5 @@
+goog.provide('my_lib.core')
+
+my_lib.core = {
+  add: function(x, y) { return x + y; }
+}